feat: Websocket server using Socket.io for instant feedback. Follows and Likes now send instant feedback to the user. note: None of this is implemented into the frontend yet.

+++ backend/src/services/user.service.ts
@@ -37,8 +37,6 @@ export class UserService {
     return jwt.sign(payload, secret, { expiresIn: '6h' });
   }
 
-  
-
 
   async register(userData: Partial<IUser>): Promise<{ user: IUser; token: string }> {
     try {
@@ -77,10 +75,8 @@ export class UserService {
   }
 
   async updateAvatar(userId: string, file: Buffer): Promise<void> {
-    const session = await this.userModel.startSession();
-    session.startTransaction();
-
     try {
+      await this.unitOfWork.executeInTransaction(async (session) => {
       const user = await this.userRepository.findById(userId, session);
       if (!user) {
         throw createError('NotFoundError', 'User not found');
@@ -94,21 +90,16 @@ export class UserService {
       if (oldAvatarUrl) {
         await this.cloudinaryService.deleteAssetByUrl(userId, oldAvatarUrl);
       }
-
-      await session.commitTransaction();
+    })
     } catch (error) {
-      await session.abortTransaction();
       throw createError(error.name, error.message);
-    } finally {
-      session.endSession();
-    }
+    } 
   }
 
+  //TODO: Finish this
   async updateCover(userId: string, file: Buffer): Promise<void> {
-    const session = await this.userModel.startSession();
-    session.startTransaction();
-
     try {
+      await this.unitOfWork.executeInTransaction(async (session) => {
       const user = await this.userRepository.findById(userId, session);
       if (!user) {
         throw createError('NotFoundError', 'User not found');
@@ -122,46 +113,42 @@ export class UserService {
       if (oldCoverUrl) {
         await this.cloudinaryService.deleteAssetByUrl(userId, oldCoverUrl);
       }
-
-      await session.commitTransaction();
+    })
     } catch (error) {
-      await session.abortTransaction();
       throw createError(error.name, error.message);
-    } finally {
-      session.endSession();
-    }
+    } 
   }
 
   async deleteUser(id: string): Promise<void> {
-    const session = await this.userModel.startSession();
-    session.startTransaction();
-
     try {
+      await this.unitOfWork.executeInTransaction(async (session) => {   
       const user = await this.userRepository.findById(id, session);
       if (!user) {
         throw createError('NotFoundError', 'User not found');
       }
 
-      const cloudResult = await this.cloudinaryService.deleteMany(user.username);
-      if (cloudResult.result !== 'ok') {
-        throw createError('CloudinaryError', cloudResult.message || 'Error deleting cloudinary data');
+      if(user.images.length > 0 ){
+        const cloudResult = await this.cloudinaryService.deleteMany(user.username);
+        if (cloudResult.result !== 'ok' ) {
+          throw createError('CloudinaryError', cloudResult.message || 'Error deleting cloudinary data');
+        }
       }
+     
 
       await this.imageRepository.deleteMany(id, session);
       await this.userRepository.delete(id, session);
-      await session.commitTransaction();
+    })
     } catch (error) {
-      await session.abortTransaction();
+    
       throw createError(error.name, error.message);
-    } finally {
-      session.endSession();
-    }
+    } 
+  
   }
 
   async getUserById(id: string): Promise<IUser> {
     const user = await this.userRepository.findById(id);
     if (!user) {
-      throw createError('NotFoundError', 'User not found');
+      throw createError('PathError', 'User not found');
     }
     return user;
   }
@@ -171,75 +158,105 @@ export class UserService {
   }
 
   async likeAction(userId: string, imageId: string): Promise<void> {
+  try {
+   
+
+    await this.unitOfWork.executeInTransaction(async (session) => {
+      console.log(`running in this.unitOfWork.executeInTransaction`);
+       // Check if image exists before starting transaction
+    const existingImage = await this.imageRepository.findById(imageId);
+    console.log(existingImage)
+    if (!existingImage) {
+      throw createError('PathError', `Image with id ${imageId} not found`);
+    }
+    
+      const existingLike = await this.likeRepository.findByUserAndImage(userId, imageId, session);
+      
+      if (existingLike) {
+        // Unlike flow
+        await this.likeRepository.deleteLike(userId, imageId, session);
+        await this.imageRepository.findOneAndUpdate(
+          { _id: imageId },
+          { $inc: { likes: -1 } },
+          session
+        );
+        await this.userActionRepository.logAction(userId, "unlike", imageId, session);
+    
+      } else {
+        // Like flow
+        const userIdObject = convertToObjectId(userId);
+        const imageIdObject = convertToObjectId(imageId);
+        
+        await this.likeRepository.create({ userId: userIdObject, imageId: imageIdObject }, session);
+        await this.imageRepository.findOneAndUpdate(
+          { _id: imageId },
+          { $inc: { likes: 1 } },
+          session
+        );
+        await this.userActionRepository.logAction(userId, "like", imageId, session);
+        await this.notificationService.createNotification(
+          {
+            receiverId: existingImage.user.id.toString(),
+            actionType: "like",
+            actorId: userId,
+            targetId: imageId,
+            session
+          },
+        );
+      }
+    });
+  } catch (error) {
+    if (error.name === 'PathError') {
+      throw error; 
+    }
+    throw createError('TransactionError', error.message, {
+      function: 'likeAction',
+      additionalInfo: 'Transaction failed',
+      originalError: error
+    });
+  }
+}
+
+  
+  async followAction(followerId: string, followeeId: string): Promise<void> {
+  
     try {
       await this.unitOfWork.executeInTransaction(async (session) => {
-        const existingLike = await this.likeRepository.findByUserAndImage(userId, imageId, session);
-        
-        if (existingLike) {
-          // Unlike: Remove the like
-          await this.likeRepository.deleteLike(userId, imageId, session);
-          // Decrement likes count on the image
-          await this.imageRepository.findOneAndUpdate(
-            { _id: imageId },
-            { $inc: { likes: -1 } },
-            session
-          );
-          // Log the user action
-          await this.userActionRepository.logAction(userId, "unlike", imageId, session);
+        const isFollowing = await this.followRepository.isFollowing(followerId, followeeId);
+    
+        if (isFollowing) {
+          // Unfollow logic
+          await this.followRepository.removeFollow(followerId, followeeId, session);
+          await this.userRepository.update(followerId, { $pull: { following: followeeId } }, session);
+          await this.userRepository.update(followeeId, { $pull: { followers: followerId } }, session);
+          await this.userActionRepository.logAction(followerId, "unfollow", followeeId, session);
         } else {
-          // Like: Create a new like
+          // Follow logic
+          await this.followRepository.addFollow(followerId, followeeId, session);
+          await this.userRepository.update(followerId, { $addToSet: { following: followeeId } }, session);
+          await this.userRepository.update(followeeId, { $addToSet: { followers: followerId } }, session);
+          await this.userActionRepository.logAction(followerId, "follow", followeeId, session);
+    
+
           
-          //Convert strings to mongoose.Types.ObjectId because .create expects
-          // Partial<T> aka Partial<ILike> which requires the mongoose object ID 
-          const userIdObject = convertToObjectId(userId);
-          const imageIdObject = convertToObjectId(imageId);
-          
-          await this.likeRepository.create({ userId: userIdObject, imageId: imageIdObject }, session);
-          
-          // Incermet the likes count
-          await this.imageRepository.findOneAndUpdate(
-            { _id: imageId },
-            { $inc: { likes: 1 } },
-            session
+          // for now I'll emit the websocket event inside the transaction
+          await this.notificationService.createNotification(
+            {
+              receiverId: followeeId,
+              actionType: "follow",
+              actorId: followerId,
+              session
+            },
+             // Pass transaction session
           );
-          // Log the user action
-          await this.userActionRepository.logAction(userId, "like", imageId, session);
+          
         }
       });
     } catch (error) {
-      throw createError(error.name, error.message, {
-        function: 'likeAction',
-        additionalInfo: 'Transaction failed'
-      });
+      throw error;
     }
+    
   }
 
-  //TODO
-  // async toggleFollowAction(followerId: string, followeeId: string): Promise<void> {
-   
-  
-  //   try {
-  //     const isFollowing = await this.followRepository.isFollowing(followerId, followeeId);
-  
-  //     if (isFollowing) {
-  //       // Unfollow Logic
-  //       await this.followRepository.removeFollow(followerId, followeeId, session);
-  //       await this.userRepository.update(followerId, { $pull: { following: followeeId } }, session);
-  //       await this.userRepository.update(followeeId, { $pull: { followers: followerId } }, session);
-  //       await this.userActionRepository.logAction(followerId, "unfollow", followeeId);
-  //     } else {
-  //       // Follow Logic
-  //       await this.followRepository.addFollow(followerId, followeeId, session);
-  //       await this.userRepository.update(followerId, { $addToSet: { following: followeeId } }, session);
-  //       await this.userRepository.update(followeeId, { $addToSet: { followers: followerId } }, session);
-  //       await this.notificationService.createNotification({ userId: followeeId, actionType: "follow", actorId: followerId });
-  //       await this.userActionRepository.logAction(followerId, "follow", followeeId);
-  //     }
-  
-  //   } catch (error) {
-
-  //   } 
-  // }
-  
   
 }
